Create homework 2
The last Lecture 



I found the last lecture to contain a lot of material stated by Randy to allow and welcome a great deal of perseverance and inspiration to be flooded into our mindsets. Going through his childhood dreams was a great way to go about this lecture for that is something nearly every individual on earth has or once had and although that could be used as a way to say something like your dreams are to late and may night come to fruition but yet its as if he's showing you ways that you can now apply to your life moving forward to achieve the goals of the oncoming future no matter the age of the listener. I found the idea of how you say something to someone and word choosing to get the same message across but with different reactions and intentions to be very interesting and made me evaluate how in many situations I may tend to react in a negative light when it could be completely avoidable with my word choice and decision to present myself through to those around me. “If you didn’t get what you want, you’ll still have the experience” was one of the many great quotes that I heard in the lecture that has certainly resonated with me and I will not be forgetting any time soon, it made me go back and remember the concerts i've played or the situations i've endured with spite and understand that there's a bright side to it all and that is the growth and experience from the moments that may be deemed as mistakes or unfortunate. Lastly I found his approach to certain dreams that may not exactly be obtainable to be very humbling with the idea of searching for the next best thing and realizing the significance of that rather than losing yourself in the original dream you so desperately wanted even though it may be impossible